refactor(amf-deserializer): restructured for more clarity * also added better docstrings
